+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ The format is based on `Keep a Changelog <https://keepachangelog.com>`_.
 - Add ``Polyclonal.mut_icXX_df`` method
 - Add ``Polyclonal.mut_icXX_plot`` method
 - Add ``PolyclonalCollection.mut_icXX_df_replicates``, ``PolyclonalCollection.mut_icXX_df``, and ``PolyclonalCollection.mut_icXX_df_w_model_values`` methods.
+- Add ``PolyclonalCollection.mut_icXX_plot`` method
 
 5.3
 ---